Fixing round off error with calculating dx in geo grids

+++ writenetcdf.cpp
@@ -1027,9 +1027,9 @@ void readgridncsize(const std::string ncfile, int &nx, int &ny, int &nt, double 
 
 	}
 
-	float dxx;
+	double dxx;
 	//check dx
-	dxx = (float) abs(xcoord[0] - xcoord[nx - 1]) / (nx - 1);
+	dxx = abs(xcoord[0] - xcoord[nx - 1]) / (nx - 1.0);
 	//dyy = (float) abs(ycoord[0] - ycoord[(ny - 1)*nx]) / (ny - 1);
